Move VisualizationRunningState down into visualizations
The loading overlay is rendering, not query editing: it lived in
querying only by colocation with QueryVisualization, and visualizations
reached up to querying for it. It now lives beside Visualization;
querying imports it downward. Its spec moves with it (the file under
query_builder only tested this component).
