+++ frontend/src/metabase/querying/components/QueryVisualization/QueryVisualization.tsx
@@ -1,17 +1,15 @@
 import cx from "classnames";
 import type { ReactNode } from "react";
 import { useEffect, useState } from "react";
-import { useTimeout } from "react-use";
 import { c, t } from "ttag";
 
 import EmptyCodeResult from "assets/img/empty-states/code.svg";
 import { Warnings } from "metabase/common/components/Warnings";
 import QueryBuilderS from "metabase/css/query_builder.module.css";
-import { useSelector } from "metabase/redux";
-import { getWhiteLabeledLoadingMessageFactory } from "metabase/selectors/whitelabel";
-import { Box, Flex, Loader, Stack, Text, Title } from "metabase/ui";
+import { Box, Flex, Stack, Text } from "metabase/ui";
 import { isMac } from "metabase/utils/browser";
 import { SERVER_ERROR_TYPES } from "metabase/utils/errors";
+import { VisualizationRunningState } from "metabase/visualizations/components/Visualization/VisualizationRunningState";
 import { prefetchVisualizationComponent } from "metabase/viz-core";
 import * as Lib from "metabase-lib";
 import { HARD_ROW_LIMIT } from "metabase-lib/v1/queries/utils";
@@ -21,8 +19,6 @@ import { RunButtonWithTooltip } from "./RunButtonWithTooltip";
 import { VisualizationError } from "./VisualizationError";
 import { VisualizationResult } from "./VisualizationResult";
 import type { QueryVisualizationProps } from "./types";
-
-const SLOW_MESSAGE_TIMEOUT = 4000;
 
 export function QueryVisualization(props: QueryVisualizationProps) {
   const {
@@ -121,35 +117,6 @@ const VisualizationEmptyState = ({ children }: { children: ReactNode }) => {
     </Flex>
   );
 };
-
-export function VisualizationRunningState({
-  className = "",
-}: {
-  className?: string;
-}) {
-  const [isSlow] = useTimeout(SLOW_MESSAGE_TIMEOUT);
-
-  const getLoadingMessage = useSelector(getWhiteLabeledLoadingMessageFactory);
-
-  // show the slower loading message only when the loadingMessage is
-  // not customized
-  const message = getLoadingMessage(isSlow() ?? false);
-
-  return (
-    <Flex
-      className={cx(className, QueryBuilderS.Overlay)}
-      c="core-brand"
-      direction="column"
-      justify="center"
-      align="center"
-    >
-      <Loader size="lg" />
-      <Title c="core-brand" order={3} mt="lg">
-        {message}
-      </Title>
-    </Flex>
-  );
-}
 
 type VisualizationDirtyStateProps = {
   className?: string;
