+++ frontend/src/metabase/querying/components/QueryVisualization/index.ts
@@ -1,6 +1,5 @@
 export {
   QueryVisualization,
-  VisualizationRunningState,
   VisualizationDirtyState,
 } from "./QueryVisualization";
 export type { QueryVisualizationProps } from "./types";
